services/persistence-service/src/requirements.txt
confluent-kafka==1.9.2
loguru==0.7.3
influxdb-client==1.36.0
watchdog==6.0.0
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-exporter-otlp==1.29.0
opentelemetry-instrumentation==0.50b0